Fix Issue + Add important info to file
